Added non-negative stocks
Added a flag to stocks to mark them as non-negative, that means that stocks can not hold a negative value. Flows can not drain a stock when the stock is empty. Added a logic in the simulation to respect this constraint.

Current implementation assigns the output to the flows arbitrarily (roughly corresponding to the order in which the flows were added, but that should not be assumed). Explicit prioritisation of flows needs to be added.

- Added Stock.allowsNegative: Bool with default value of false - we expect most of the models using non-negative stocks
- Stock: Exposed the new flag through attributes
- changed lookup in SimulationState to be node-based not name-based
- Simulation.last renamed to current
- added logic to Simulation.step() to account for non-negative nodes
- added test for non-negativity of stocks

Housekeeping:

- moved ParameterSelector from ExpressionNode to Model so we have all graph metadata in one place